+++ clang/utils/TableGen/ClangOptionDocEmitter.cpp
@@ -40,25 +40,23 @@ struct DocumentedGroup : Documentation {
   Record *Group;
 };
 
-bool hasFlag(const Record *Option, StringRef OptionFlag) {
-  for (const Record *Flag : Option->getValueAsListOfDefs("Flags"))
+bool hasFlag(const Record *Option, StringRef OptionFlag, StringRef FlagsField) {
+  for (const Record *Flag : Option->getValueAsListOfDefs(FlagsField))
     if (Flag->getName() == OptionFlag)
       return true;
   if (const DefInit *DI = dyn_cast<DefInit>(Option->getValueInit("Group")))
-    for (const Record *Flag : DI->getDef()->getValueAsListOfDefs("Flags"))
+    for (const Record *Flag : DI->getDef()->getValueAsListOfDefs(FlagsField))
       if (Flag->getName() == OptionFlag)
         return true;
   return false;
 }
 
 bool isOptionVisible(const Record *Option, const Record *DocInfo) {
-  for (StringRef Exclusion : DocInfo->getValueAsListOfStrings("ExcludedFlags"))
-    if (hasFlag(Option, Exclusion))
+  for (StringRef IgnoredFlag : DocInfo->getValueAsListOfStrings("IgnoreFlags"))
+    if (hasFlag(Option, IgnoredFlag, "Flags"))
       return false;
-  if (!DocInfo->getValue("IncludedFlags"))
-    return true;
-  for (StringRef Inclusion : DocInfo->getValueAsListOfStrings("IncludedFlags"))
-    if (hasFlag(Option, Inclusion))
+  for (StringRef Vis : DocInfo->getValueAsListOfStrings("VisibilityMask"))
+    if (hasFlag(Option, Vis, "Vis"))
       return true;
   return false;
 }